backend/requirements.txt
flask
pillow
google-generativeai
